move cmd to internal package. fixes #166
